+++ cli/file_fetcher.rs
@@ -18,10 +18,10 @@ use deno_core::futures;
 use deno_core::futures::future::FutureExt;
 use deno_core::parking_lot::Mutex;
 use deno_core::ModuleSpecifier;
+use deno_runtime::deno_fetch::create_http_client;
 use deno_runtime::deno_fetch::reqwest;
 use deno_runtime::deno_web::BlobStore;
 use deno_runtime::permissions::Permissions;
-use deno_tls::create_http_client;
 use deno_tls::rustls::RootCertStore;
 use log::debug;
 use log::info;
